docker/fastapi_postgresql_test/fastapi/requirements.txt
fastapi
python-multipart
python-dotenv
pandas
numpy
pydantic-settings
sqlmodel
uvicorn
psycopg2-binary
sqlalchemy
